server/ai_core_service/requirements.txt
# Core Web Framework & Server
Flask==3.0.2
Flask-CORS==4.0.0
waitress==3.0.0  # Production WSGI server

# Configuration & Utilities
python-dotenv==1.0.1
uuid==1.30  # Standard library, listed for explicitness

# AI - Langchain Ecosystem
langchain==0.1.12
langchain-community==0.0.27
langchain-huggingface==0.0.10
langchain-ollama==0.0.3

# AI - LLM & Embeddings Libraries
ollama==0.1.6
sentence-transformers==2.5.1

# AI - Vector Store
faiss-cpu==1.7.4  # CPU version of FAISS

# Document Processing & Text Extraction
pypdf==4.1.0
PyPDF2==3.0.1
pymupdf==1.23.26
python-docx==1.1.0
python-pptx==0.6.23

# Tokenizers
tiktoken==0.6.0

# HTTP Client Libraries
requests==2.31.0
httpx==0.27.0

# LLM Providers
google-generativeai==0.3.2
groq==0.4.2

# Logging and Monitoring
logging==0.4.9.6